renamed test to examples
